Work-around for conflicting packages on linux-modern.
Packages conflict. The same commit worked fine the previous day, so it's likely not our fault. We should probably inform maintainers, but until then, try to work around the problem without a blanket ignore of all 'overwrite' type errors. The only clear thing is that this all has to do with 'odbc', which is database related. We use that to cache certain things on startup, so likely we actually need it.

part of CURA-10290